refactor: change styles file name
